backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.4.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2 